refactor api helpers for readability/scalability
