+++ frontend/helpers/api/config.ts
@@ -1,0 +1,1 @@
+export const baseURL = process.env.NEXT_PUBLIC_API_BASE_URL || "http://localhost:5000/api/v1"
