+++ frontend/helpers/api/index.ts
@@ -1,2 +1,5 @@
-export * from "./api"
+export * from "./base"
+export * from "./config"
+export * from "./auth"
+export * from "./incidents"
 export { useMockServiceWorker, apiMode } from "./mocks/browser.setup"
